Rename and refactor macros to improve package structure
Moved macros from `pkg/repo/macro` to `pkg/core/command` for better organization. Updated related function names and types to align with the new structure, including `NewMacroTemplates` renamed to `NewMacro`. Adjusted code to ensure compatibility with the changes, improving maintainability.
